requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.4
Django==6.0.2
django-environ==0.13.0
django-filter==25.2
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0

